# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/outlets": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return list of outlet",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Get all outlet",
                "parameters": [
                    {
                        "type": "string",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search_fields",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sort_dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.IPagination-array_models_Outlets"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save outlet data in Db.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Create Outlet",
                "parameters": [
                    {
                        "description": "Create outlet",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateOrUpdateOutlet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Outlets"
                        }
                    }
                }
            }
        },
        "/outlets/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the outlet whoes OutletId value mathes id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Get Single outlet by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "get outlet by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Outlets"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update outlet data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Update outlet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "update outlet by id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update outlet",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateOrUpdateOutlet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Outlets"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return the outlet whoes outletId valu mathes id.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Outlets"
                ],
                "summary": "Get Single outlet by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "delete outlet",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Outlets"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Outlets": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "meta_data": {
                    "type": "string"
                },
                "outlet_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
        "requests.CreateOrUpdateOutlet": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "meta_data": {
                    "type": "string"
                },
                "outlet_name": {
                    "type": "string"
                },
                "warehouse_id": {
                    "type": "string"
                }
            }
        },
...
        "utils.IPagination-array_models_Outlets": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Outlets"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "meta": {},
                "page": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}